adding notes and moving test to zk module
