+++ crates/forge/tests/cli/script.rs
@@ -7,8 +7,8 @@ use forge_script_sequence::ScriptSequence;
 use foundry_test_utils::{
     rpc::{self, next_http_rpc_endpoint},
     snapbox::IntoData,
-    util::{OutputExt, OTHER_SOLC_VERSION, SOLC_VERSION},
-    ScriptOutcome, ScriptTester, ZkSyncNode,
+    util::{OTHER_SOLC_VERSION, SOLC_VERSION},
+    ScriptOutcome, ScriptTester,
 };
 use regex::Regex;
 use serde_json::Value;
@@ -2174,117 +2174,6 @@ forgetest_async!(can_deploy_library_create2_different_sender, |prj, cmd| {
         .broadcast(ScriptOutcome::OkBroadcast)
         .assert_nonce_increment(&[(2, 2)])
         .await;
-});
-
-//         vm.zkVm(True);
-
-// <https://github.com/foundry-rs/foundry/issues/8993>
-forgetest_async!(test_zk_broadcast_raw_create2_deployer, |prj, cmd| {
-    foundry_test_utils::util::initialize(prj.root());
-    let node = ZkSyncNode::start().await;
-    let url = node.url();
-
-    println!("1. url: {:?}", url);
-
-    let (_, private_key) = ZkSyncNode::rich_wallets()
-        .next()
-        .map(|(addr, pk, _)| (addr, pk))
-        .expect("No rich wallets available");
-
-    //print private key
-    println!("1. private_key: {:?}", private_key);
-
-    prj.add_source("Counter.sol", r#"
-    pragma solidity ^0.8.0;
-
-    contract Counter {
-        uint256 public count;
-        function increment() external {
-            count++;
-        }
-    }
-    "#).unwrap();
-
-    //deploy
-    let output = cmd.args([
-        "create",
-        "src/Counter.sol:Counter",
-        "--zksync",
-        "--private-key",
-        private_key,
-        "--rpc-url",
-        &url,
-    ]);
-    let address = output.assert_success().get_output().stdout_lossy();
-    println!("2. address: {:?}", address.trim());
-    cmd.forge_fuse();
-    // println!("2. private_key: {:?}", private_key);
-
-    // let output = cmd
-    //     .cast_fuse()
-    //     .args([
-    //         "mktx",
-    //         "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
-    //         "deposit()",
-    //         "--rpc-url",
-    //         &url,
-    //         "--private-key",
-    //         private_key,
-    //         "--zksync",
-    //     ])
-    //     .assert_success()
-    //     .get_output()
-    //     .stdout_lossy();
-
-    // println!("3. output: {:?}", output);
-
-    prj.add_script(
-        "Foo",
-        r#"
-import "forge-std/Script.sol";
-import {Counter} from "../src/Counter.sol";
-contract SimpleScript is Script {
-    function run() external {
-        // zk raw transaction
-        vm.startBroadcast();
-        // This raw transaction comes from cast mktx of increment() to Counter contract
-        vm.broadcastRawTransaction(
-            hex"71f88580808402b275d08304d718949c1a3d7c98dbf89c7f5d167f2219c29c2fe775a78084d09de08a80a0f76c089059f46bb90adc0b34fa643edf175413b9e076185f46afe84f9283ccfba077b2a9878429569852f9e114e0c2bd8e67be25048752c3b2acab6cd7e2cdf4ff82010494bc989fde9e54cad2ab4392af6df60f04873a033a80c08080"
-        );
-        vm.stopBroadcast();
-    }
-}
-"#,
-    )
-    .unwrap();
-
-    println!("4. prj.root(): {:?}", prj.root());
-
-    cmd.args([
-        "script",
-        "--zksync",
-        "--private-key",
-        private_key,
-        "--rpc-url",
-        &url,
-        "--broadcast",
-        "--slow",
-        "--non-interactive",
-        "SimpleScript",
-    ]);
-
-    println!("5. cmd: {:?}", "ok");
-
-    let output = cmd.assert_success().get_output().stdout_lossy();
-    println!("6. output: {:?}", output);
-
-    assert!(output.contains("ONCHAIN EXECUTION COMPLETE & SUCCESSFUL."));
-
-    // assert!(!api
-    //     .get_code(address!("4e59b44847b379578588920cA78FbF26c0B4956C"), Default::default())
-    //     .await
-    //     .unwrap()
-    //     .is_empty());
 });
 
 // <https://github.com/foundry-rs/foundry/issues/8993>
